moved tab completion to separate file, updated handleTab() to new structure.  Created new global to hold a copy of the tab completion slice thats always up to date instead of generating each time, hooked completion options updates into view updates and channel join activity.
